Worked on the giphy comment feature

+++ Tweeter/resources/views/tweeter_tweet_view.blade.php
@@ -52,10 +52,15 @@
                                 <p class="control">
                                 </p>
                             </div>
-                            <div class="field">
-                                <p class="control">
-                                <button class="button is-link" type="submit">Post comment</button>
-                                </p>
+                            <div class="level">
+                                <div class="level-item">
+                                    <p class="control">
+                                    <button class="button is-link" type="submit">Post comment</button>
+                                    </p>
+                                </div>
+                                <div class="level-item">
+                                <giphy tweet-id="{{$tweet->id}}"/>
+                                </div>
                             </div>
                         </form>
                     </div>
@@ -65,7 +70,7 @@
             @foreach ($tweet->comments as $comment)
             <article class="media">
                 <figure class="media-left">
-                    <p class="image is-48x48">
+                    <p class="image is-32x32">
                         <a href="/profile/view/{{$comment->users->id}}">
                       <img src="/storage/avatars/{{$comment->users->profiles->avatar}}"></a>
                     </p>
@@ -76,22 +81,38 @@
                       <a class="has-text-primary is-size-6" href="/profile/view/{{$comment->users->id}}"><strong>{{$comment->users->profiles->firstname}} {{$comment->users->profiles->lastname}}</strong></a>
                       <a class="has-text-primary is-size-6" href="/profile/view/{{$comment->users->id}}"><small>@ {{$comment->users->username}}</small></a>
                         <br>
-                        {{$comment->content}}
+                       @php
+                           if (filter_var($comment->content, FILTER_VALIDATE_URL)) {
+                                echo "<img src='$comment->content'>";
+                            } else {
+                                echo $comment->content;
+                            }
+                       @endphp
                         <br>
                         <small>{{$comment->created_at->diffForHumans()}}</small>
                       </p>
                     </div>
                 </div>
-                <div class="media-right">
+                <div class="">
+                    <div class="level"></div>
+
                     @if ($comment->user_id == Auth::user()->id)
-                    <a class="button is-warning is-small" href="/tweet/comment/edit?commentId={{$comment->id}}&tweetId={{$tweet->id}}">
-                        <span class="icon is-small"><i class="fas fa-edit"></i></span>
-                        <span>Edit</span>
-                    </a>
-                    <a class="button is-danger is-small" href="/tweet/comment/delete?commentId={{$comment->id}}">
-                        <span class="icon is-small"><i class="fas fa-trash"></i></span>
-                        <span>Delete</span>
-                    </a>
+                        @if (filter_var($comment->content, FILTER_VALIDATE_URL))
+                            <a class="button is-danger is-small" href="/tweet/comment/delete?commentId={{$comment->id}}">
+                                <span class="icon is-small"><i class="fas fa-trash"></i></span>
+                                <span>Delete</span>
+                            </a>
+                        @else
+                            <a class="button is-warning is-small" href="/tweet/comment/edit?commentId={{$comment->id}}&tweetId={{$tweet->id}}">
+                                <span class="icon is-small"><i class="fas fa-edit"></i></span>
+                                <span>Edit</span>
+                            </a>
+                            <a class="button is-danger is-small" href="/tweet/comment/delete?commentId={{$comment->id}}">
+                                <span class="icon is-small"><i class="fas fa-trash"></i></span>
+                                <span>Delete</span>
+                            </a>
+                        @endif
+
                 @endif
                 </div>
             </article>
